% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyMultiple.R
\name{FuzzyDouble}
\alias{FuzzyDouble}
\title{FuzzyDouble}
\usage{
FuzzyDouble(x)
}
\arguments{
\item{x}{Number}
}
\value{
The double of x with fuzz
}
\description{
FuzzyDouble
}
\examples{
FuzzyDouble(2)
4.023
}
