% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyMultiple.R
\name{FuzzyMultiple}
\alias{FuzzyMultiple}
\title{FuzzyMultiple}
\usage{
FuzzyMultiple(x, n)
}
\arguments{
\item{x}{Number}

\item{n}{Number 2}
}
\value{
x times n with fuzz
}
\description{
FuzzyMultiple
}
\examples{
FuzzyMultiple(2,3)
6.0321
}
