% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyMultiple.R
\name{FuzzyTriple}
\alias{FuzzyTriple}
\title{FuzzyTriple}
\usage{
FuzzyTriple(x)
}
\arguments{
\item{x}{Number}
}
\value{
The triple of x with fuzz
}
\description{
FuzzyTriple
}
\examples{
FuzzyTriple(2)
6.0422
}
